M4-5: React Context - Exercise 2
